Sorted neighbor communication ordering so no races will occur
I came up with an algorithm for sorting the neighbor
communications. It was tested in the algorithms python script. Now I
can rearrange the hill exchange code to be simpler as well.

+++ lib/gaussian_grid.h
@@ -176,7 +176,7 @@ class DimmedGaussGrid : public GaussGrid{
       index1 = i - minisize_total_ / 2;
       for(j = 0; j < DIM-1; j++) {
 	index[j] = index1 % minisize_[j];
-	index1 = (index1 - index[j]) / minisize_[j];
+	index1 = (index1 - index[j]) / static_cast<long int>(minisize_[j]);
       }
       index[j] = index1;
